Started basic account system & added base template.
Just a basic login / registration system to make development a bit easier.
Used the default login app from Django to handle the login view, but it is very simple to switch to a custom view in case we want to handle it another way.

In total added 3 views:
- login
- signup
- account

Base template in base.html

Issues: #9 #10