Added extra user info model & admin page control
The model UserExtraInfo in account/models.py has been created. It's linked to a User and can be accessed through user.userextrainfo in both Python as well as the templates (through the request.user object).

I've also registered the model with Django's admin system (see line 11 in account/admin.py). This allows an admin user to view & edit the information in the model.

The new account_page.html is an example on how to access said extra information or use a default value if the extra information doesn't exist.