Removed unnecessary user id field for likes & wishlist lists. Fixed issue with double json loading of data.
User id is already stored once in the json object at the top level, no need to include it for each liked ship / wishlisted ship item.

Also fixed a bug where the serializer already loaded it as a valid json string, and the JsonResponse then double loaded that which caused it to be incorrectly formatted.