Added default superuser account
A superuser account with username="Admin" and password="Admin" is now created when your Django container is started. This should resolve potential problems where a developer doesn't have direct access to the container and can't set up a superuser.

Notice: we absolutely do not want this (and the other docker-compose.yml variables) to be set like this for production.