Fix for simplejwt issue 266
Users with outstanding tokens cannot be deleted if the simplejwt token blacklist app is enabled. This is because outstanding tokens override admin permissions. This fix overrides that and gives admins permission to delete outstanding tokens (and as such; permission to delete users with outstanding tokens).

More info: https://github.com/jazzband/djangorestframework-simplejwt/issues/266